iterator: fix panic when retry policy returns IgnoreWriteError in iterator
Currently, the RowIterator expects that its background worker will
always send at least one page (or an error) to the iterator. However,
there is an edge case where it may not be true: if the first request
executed via query_iter/execute_iter fails and the retry policy returns
an IgnoreWriteError decision (currently returned only by downgrading
consistency policy in case of WriteTimeout). In such case, those methods
will trigger a panic.

This commit fixes the problem by returning a dummy, empty page to the
iterator in that case.