session_builder: made builder generic over mode
With addition of cloud serverless support, we want to make a clear
separation between construction of Sessions operating in serverless mode
and those in default mode. The reason for that is we want to have a
clean API, which prevents users from doing weird things, such as
providing both the cloud config yaml and known_nodes IPs (such
combination is inherently incompatible). Another important goal was to
leave the builder's API untouched. We managed to achieve this using
Rust's type inference, which enables the user to construct the builders
generic over their mode without even being conscious that the builder is
generic. This is done using the constructors: `new()` for the default mode
and `new_cloud(config_yaml)` for the cloud mode. Each mode restricts its
configuration options to those available in the given mode.