cluster, topology: ContactPoint added
This struct will come in handy in further parts of this refactor.
Also, for serverless cloud connections to be established, the datacenter
of the endpoint is required to be known. Therefore, it is passed along
with its address on the way from SessionConfig as well as from
MetadataReader to the place of usage (added in serverless-related
commits).