caching_session: use custom hasher instead of ahash in tests
Currently, we are importing the `ahash` crate for only one reason: to
test that CachingSession works with custom hasher implementations. The
custom hasher quality doesn't matter at all in this test, so it's very
easy to write own implementation for testing purposes, rendering the
dependency on `ahash` obsolete.

The change was prompted by a report from one of our users which use a
fork of the driver and managed to trigger a bug in cargo - most likely
this one: https://github.com/rust-lang/cargo/issues/7463 .
The most likely cause was that the project specified a different,
non-compatible version of ahash in the dependencies than the one in
scylla's dev-dependencies. While it doesn't fix the cargo bug and it can
still happen with other dev-dependencies, the bug should no longer
happen with conflicting `ahash`. Moreover, removing an unneeded
depdencency from a project is a good motivation in itself.