treewide: appease cargo doc
The new version of cargo doc started complaining about unescaped
`<` and `>` signs because they looked like html tags. This commit gets
rid of this complaint either by putting them into `backticks` if they
are a part of some rust code mentioned in a docstring, or escaping them
with a backslash in other cases.