scylla-macros: support renaming the scylla crate
Currently, the macros will only work, if the code using them has
a dependency on the scylla driver, or on the scylla_cql if it's
renamed to scylla.
This is caused by the fact that the paths referring to the scylla
code are hardcoded to use a `scylla` crate.
As a result, if a someone uses the `scylla_cql` crate without
renaming, or renames `scylla_cql` or `scylla` to a name differen
name than `scylla`, the macros will not work.
Similarily, when another crate re-exports the macros, or uses them
in their own macros, it can only work if the crate is used under
the name `scylla` and it additionally re-exports parts of the
`scylla_cql` code that are used in the macro under the same paths.

This patch adds a `scylla_crate` helper attribute that allows users to
set the crate name that will export the code that is used in the
macros. All the symbols that are used there are also additionally
exported in a `_macro_internal` module to make re-exporting important
code segments more convenient.