deserialization: make deser_cql_value public
In some cases (in particular, Scylla UDFs), the cql values are
passed in a serialized form without the context of a row.

The current implementation of deser_cql_value function is already
able to parse these values, so it can be just made public to be usable
in those cases.