Revert "topology: Added address translation on peer query"
This reverts commit 9bf4f74f421faf76122b30c78ec61299fba1783a.

In the following commits in this PR, address translation will be moved
to PoolRefiller upon opening connection. The reason for that is the need
of having PoolRefillers with not-yet-resolved domain names. This will
allow supporting nodes (and serverless clusters) that are accessible
through a number of IPs. Especially, when a subset of those IPs is
unavailable, other ones will be able to be used for establishing the
connection successfully.