cluster: known_peers indexed by host_id instead of IP
We need to have known_peers indexed by host_id for node identification (if host_id changes, we cannot recycle the Node struct). The bad is that we have to to iterate through the whole known_peers upon node down marker change, but there will rarely more than ~100 nodes, and changes of down marker are infrequent enough to afford this.
As an important tradeoff, we still keep only one hashmap of known_peers.