StatementConfig: serial_consistency as double Option
We need to differentiate between serial consistency unset in the frame
(or, in other words, set to None) vs set to inherit its value from
some parent configuration (cascading behaviour). Therefore, the new
semantics is as follows:
None -> represents that serial consistency is inherited from the parent configuration (or unset, if there is no parent configuration),
Some(None) -> represents that serial consistency is set to ANY (or, in other words, it is unset),
Some(Some(serial_consistency) -> represents all remaining set serial consistencies (that is, Serial and LocalSerial).