cql_to_rust: impl FromCqlValue for Date
The driver supports serializing and deserializing between the `date` CQL
type and the `chrono::NaiveDate` type. However, the `NaiveDate` type is
not always sufficient as it can't represent the whole range of values
the `date` CQL type can. For that purpose, we provide a `Date` type that
is able to represent all CQL values of `date` type.

Until now, it was only possible to serialize `Date`. This commit adds a
missing FromCqlValue impl necessary for deserialization.