CI/Makefile: disable `uninlined_format_args` clippy lint
New version of clippy introduced the `uninlined_format_args` lint which
recommends putting variable names directly into format strings, like
this:

    let what = "world";
    println!("Hello, {what}!");

    // Instead of:
    println!("Hello, {}!", what);

The problem with this feature is that rust-analyzer does not support
renaming variables inside format strings yet. The lint has been
downgraded to pedantic level and will most likely appear as such in the
next release.

This commit disables the new lint in the CI and local Makefile. We can
turn it on again after this feature gets better support from
rust-analyzer.