proxy: Ports are assigned from private/dynamic pool
When attempting to connect to shard aware port, proxy should respect the
port number of the driver. As it assigned ports that do not belong to
the private pool (49152-65525), bind() failed silently (returning Ok),
so the socket was not bound. Hence, it was assigned a random port.
This commit fixes it by reusing code from the driver that ensures that
only ports from the mentioned pool are considered.