execution_profile: Added integration test for profiles
What's tested? Whether set execution profiles correctly influence the
queries being sent. We check the precedence:
session profile < statement profile < statement options
by testing that correct Consistency, Load balancing policy and retry
policy is used. The weak point is that we do not explicitly test
SerialConsistency, SpeculativeExecutionPolicy and request timeouts, but
as they come as a package-deal analogous to the tested features, we can
safely assume they work as well.
The test uses proxy in order to forge errors to trigger RetrySession, as
then it receives Consistency of the query as a parameter and this way we
can assert that the consistency is correct.