topology: use a structure to represent a row from system.{local,peers}
Thanks to #628, it is now possible to provide an explicit path to a
crate which provides necessary items for the FromRow macro. It is now
possible to use this macro from the `scylla` crate itself, as we just
need to point it to the `scylla-cql` crate instead.

The NodeInfoRow struct is introduced to represent rows fetched from
system.local and system.peers. It looks nicer than using a 5-element
tuple and will reduce some noise when performing refactors in the
commits that follow.