topology: convert query_peers to use query_iter instead of query_all
Previously, query_peers used query_all to fetch topology information and
then processed it using iterators. Now, this method is changed to use
query_iter and rows are processed on the fly using asynchronous streams.

In order to preserve the behavior of system.peers and system.local being
queried in parallel, stream::select is used which merges results of both
streams as they become available. Previously, we just put results of one
query after results of the other one. Because of this, the order of
Peers in the final result might be slightly different on each fetch -
however, this order was never specified and could actually change e.g.
when the control connection changes, so I think this should be OK.