connection: get rid of execute_all
This method is a sibling of query_all that was never actually used. In
the future we will consider switching to prepared statements for
internal queries, but 1) this will be done via not-yet-existing
execute_iter and 2) we might do it by converting query_iter to use
prepared statements. Therefore, there is no need to keep the dead code.

This commit gets rid of execute_all and its all occurrences in the
connection_query_all_execute_all_test (now renamed to:
connection_query_all_test).