connection: change the query_all test to query_iter test
The query_iter method is meant to serve similar purpose to what
query_all does currently - fetch data on control connection for internal
queries - so it makes sense to adapt the (only) existing test for the
new method.

The last assertion (4) had to be removed because the new method does not
check that the page size is set by the caller.