docker-compose.yml: improve settings and boot order
Up until now, our docker-compose file that sets up Scylla cluster was
starting up all nodes in parallel. This is not something that is
supported in Scylla and it can sometimes fail. We didn't see those
failures until now because we didn't print the cluster logs. This commit
makes sure that nodes are booted up in order.

In addition to this, now nodes start with a fixed number of shards (2
each) and memory (1GB each). I'm not sure a more powerful cluster is
needed, and - by default - the shard count is equal to the number of
cores of the host machine. A large number of shards can quickly eat up
the /proc/sys/fs/aio-nr resource in the system, so the cluster wasn't
really usable on machines with a large number of cores.