token_aware: avoid iterating over the entire ring when RF is unsatisfiable
token_aware code uses cluster.ring_range(token).unique() to iterate over
candidate replicas until enough candidates are found to satisfy the RF.
This behaves badly when the number of candidates is smaller than RF --
we always iterate over the entire ring, which is very wasteful (it was
seen to slow down the driver by a factor of >2 in a simple performance
test).

Fix that by ending the iteration early when all unique candidate nodes were
already considered.

Fixes #452