topology: Added Pre{Collection,Cql}Type
After the attempted refactor, CqlType::UserDefinedType variant will
contain a fully-fledged UDT definition (UserDefinedType struct). In
order to succesfully create such variants with definitions, we need to
process the UDTs in topological order (i.e. if type A contains type B,
then B must be processed and have its definition created before A). For
the toposort to take place, we first need to build a vec of UDTs. In
order to solve the chicken-or-the-egg problem, we divide Cql types
parsing into two phases: Pre phase and main phase. In Pre phase, UDT
variant contains only the name (as it had been before). In the main
phase, Pre types are converted into final types in topological order by
enriching them with proper definitions (UserDefinedType structs).

This commit introduces the Pre types.