connection_pool: NodeConnectionPool made clonable
After a node changes its broadcast IP (which, though rarely, might
happen), or after a control connection node changes, then the node_addr
changes: either its inner IP changes (in the first case), or its enum
variant is altered (e.g. from RealPeer to FixedLocal). In such case, the
Node struct should be updated; it is, however, immutable, so a new Node
must be created. So far it led to pool recreation as well; now, we
introduce changes to preserve the pool and make it possible to update
the IP it should open new connections to.

This commit makes NodeConnectionPool clonable, which allows for
preserving the pool upon new Node creation.