connection_pool: node_addr stored in Arc<RwLock>
After a node changes its broadcast IP (which, though rarely, might
happen), or after a control connection node changes, then the node_addr
changes: either its inner IP changes (in the first case), or its enum
variant is altered (e.g. from Translatable to Untranslatable). In such
case, the Node struct should be updated; it is, however, immutable,
so a new Node must be created. So far it led to pool recreation as well;
now, we introduce changes to preserve the pool and make it possible
to update the IP it should open new connections to.

This commit puts node_addr into RwLock that is shared between
PoolRefiller and all NodeConnectionPool clones. This allows for
effective updates of node_addr in case it changes between topology
refreshes. (To be more accurate, the whole UntranslatedEndpoint is put
into Arc, but the API that the pool provides enables only node_addr
alterations).