node: Inheriting everything but address from another Node
After a node changes its broadcast IP (which, though rarely, might
happen), or after a control connection node changes, then the node_addr
changes: either its inner IP changes (in the first case), or its enum
variant is altered (e.g. from Translatable to Untranslatable). In such
case, the Node struct should be updated; it is, however, immutable,
so a new Node must be created. So far it led to pool recreation as well;
now, we introduce changes to preserve the pool and make it possible
to update the IP it should open new connections to.

This commit:
- adds a method to NodeConnectionPool that updates an endpoint that
  PoolRefiller open new connections to.
- creates a constructor of Node that clones all fields but down_marker
  (which is reset to false) and address (which is given explicitly).
  The pool is inherited as well, and it is updated with the new endpoint
  (The only intended update of the endpoint is in respect to address).
- uses the described constructor when the address of a Node changes.
  Then the old Node struct is dropped, but its pool is preserved for
  the new Node.