node: Introduced NodeAddr to signify address'es translatability
There is a known bug in older Cassandra and Scylla releases that
rpc_address in system.local is misconfigured. Thus, we replace a faulty
address found in system.local with connection's address. After such
replacement, however, address translation should definitely not be
performed on such connection's address. Thus, we cover this case
specially in NodeAddr by Untranslatable variant. This way, we distinguish
that special case from the untranslated address of a peer, which is
represented by Translatable variant.