proxy: Include shard-aware port in translation map
As address translation has been moved from topology refresh to opening
new connection, both shard-aware and non-shard-aware ports may be
provided to AddressTranslator. Therefore, the translation map that the
proxy generates automatically is extended to contain shard-aware
versions of the nodes' addresses as well.

It is important to remind the users in the next release notes that they
should adjust their translation maps as well.