prepared_statement: Made cloning cheaper by Arcs
As cloning PreparedStatements may be convenient in some workloads,
effort is done to make it possibly cheap. This commit put the heavy
heap-allocated PreparedStatement metadata and its statement String
behind an Arc, in a shared struct. In order to have cleaner inner API,
usages of data referenced in the shared struct access them now through
the getter.