Expose calculate_token
Helps (or arguably fixes) #468

For smarter batch constitution (following up on #448), it is required
to have access to the token a particular statement would be directed to.
However, that is pretty difficult to do without access to
calculate_token.

That was initially put in #508, but planned to put in a separate pr to
keep it minimal
(https://github.com/scylladb/scylla-rust-driver/pull/508#discussion_r949399700)

It seems I had forgotten to open that separate PR, and I end up getting
bitten by that now that I want to constitute my batches in a smarter
way.

So here it is.

I'm only putting "helps" above because I think we may want to also
expose query planning (
`session.plan(prepared_statement, serialized_values) -> impl Iterator<Item = (Arc<Node>, ShardID)>`
) as that may make it significantly easier - but I'd like to keep this
PR that just *enables* the ideal behavior as simple as possible.