iterator: rename (Typed)RowIterator to Legacy08(Typed)RowIterator
Like in the case of Legac08yQueryResult, TypedRowIterator will be\
replaced with a new one, and the old one preserved for compatibility
reasons. As a first step, it is renamed to Legacy08TypedRowIterator.

Similarly, RowIterator is renamed to Legacy08RowIterator to mark that it
is the part of the old API. There won't be a new RowIterator, though.