iterator: introduce poll_next_page
This commit refactors the part responsible for acquiring the next page
by the Legacy08RowIterator to a different function. This change is a
preparation necessary to support the new deserialization interface -
there will be a method that can return deserialized type that borrows
from the current iterator, and - for "lifetimes reasons" - acquiring the
next page must be put into a separate method.