transport: introduce new QueryResult
(Re)-introduces QueryResult. It is quite similar to the old
(Legacy08)QueryResult, but it keeps rows in an unparsed state (via
RawRows) and has methods that allow parsing the contents using the new
API.

Helper method names are similar to what the old QueryResult had, just
with the `_typed` suffix dropped - as now it is always required to
provide the type of rows when parsing them, this suffix sounded
redundant.