iterator: introduce ready_some_ok! macro
The iterator module has a small number of functions which return
Poll<Option<Result<T, E>>>, call other functions that return such a
stacked type and are support to continue if value obtained is not of
form Ready(Some(Ok(x))). This requires a multiline pattern match
instruction, which is basically boilerplate.

This commit introduces the ready_some_ok! macro which hides the
aforementioned boilerplate. Conceptually, it is similar to the existing
std::task::ready! macro.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>