iterator: adjust to the new deserialization framework
This commit makes the RowIteratorWorker pass raw rows to the main tokio
task, instead of the eagerly deserialized Rows.

The equivalent of the old RowIterator is now QueryPager. It cannot be
conveniently iterated on, as it does not have any information about the
column types. It features (yet not exposes) a `next()` method for
deserializing consecutive `ColumnIterator`s. Users cannot manually
perform deserialization using this method directly, because will be able
to utilise the preferred (typed) API that will be added in the next
commit. If they prefer manual deserialization, they will be be able to
specify `ColumnIterator` as the deserialized row type, and proceed with
manual deserialization from there.

The legacy iterators are preserved by wrapping around QueryPager.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>