treewide: rename QueryResult -> Legacy08QueryResult
Soon, a new QueryResult will be introduced with a slightly different
API. The old one will be preserved as Legacy08QueryResult, in order to
make migration easier.

This commit renames the existing QueryResult to Legacy08QueryResult, as
well as the query_result module to legacy_query_result. The new
QueryResult will be introduced in later commits.