result: introduce RawRows & friends
Introduces new structs for lazy deserialization of RESULT:Rows frames.

Introduces `ResultMetadataHolder`, which is, unsurprisingly, a
versatile holder for ResultMetadata. It allows 3 types of ownership:
1) borrowing it from somewhere, be it the RESULT:Rows frame
   or the cached metadata in PreparedStatement;
2) owning it after deserializing from RESULT:Rows;
3) sharing ownership of metadata cached in PreparedStatement.

Introduces new structs representing the RESULT:Rows CQL frame body,
in various phases of deserialization:
- `RawMetadataAndRawRows` only deserializes (flags and) paging size
  in order to pass it directly to the user; keeps metadata and rows
  in a serialized, unparsed form;
- `DeserializedMetadataAndRawRows` deserializes metadata, but keeps
  rows serialized;

`DeserializedMetadataAndRawRows` is lifetime-generic and can be
deserialized from `RawMetadataAndRawRows` to borrowed or owned form by
corresponding methods on `RawMetadataAndRawRows`.

`DeserializedMetadataAndRawRows` must abstract over two different ways
of storing the frame:
- shared ownership (Bytes),
- borrowing (FrameSlice).

The problem arises with the `rows_iter` method.
- in case of `DeserializedMetadataAndRawRows<RawRowsOwned>`, the struct
  itself owns the frame. Therefore, the reference to `self` should have
  the `'frame` lifetime (and this way bound the lifetime of deserialized
  items).
- in case of `DeserializedMetadataAndRawRows<RawRowsBorrowed>`, the
  struct borrows the frame with some lifetime 'frame. Therefore, the
  reference to `self` should only have the `'metadata` lifetime, as the
  frame is owned independently of Self's lifetime.

This discrepancy is not expressible by enums. Therefore, an entirely
separate `rows_iter` must be defined for both cases, and thus both cases
must be separate types - and this is guaranteed by having a different
type parameter (because they are distinct instantiations of a generic
type).

To restrict the type parameter of `DeserializedMetadataAndRawRows`
to the two valid variants (borrowed and owned), a trait `RawRowsKind`
is introduced.

Credits to Karol Baryła for replacing a macro (the first approach)
with a generic function.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>
Co-authored-by: Karol Baryła <karol.baryla@scylladb.com>