result: introduce RawRows & friends
Introduces new structs for lazy deserialization of RESULT:Rows frames.

Introduces `ResultMetadataHolder`, which is, unsurprisingly, a
versatile container for ResultMetadata. It allows 2 types of ownership:
1. owning it in a borrowed form, self-borrowed from RESULT:Rows frame;
2. sharing ownership of metadata cached in PreparedStatement.

The first type of ownership is made possible by use of the `yoke` crate,
which drives `SelfBorrowedMetadataContainer`: a container that can be
considered an `Arc<ResultMetadata>` with an additional capability of
containing metadata in a borrowed form.
The borrow comes from the `Bytes` that this container holds internally.
Therefore, the held `ResultMetadata`'s lifetime is covariant with the
lifetime of this container itself.

To make `yoke` work in our use case, the first two lines of `unsafe`
code is introduces to our driver. However, the requirements that those
unsafe traits impose on their implementees are obviously met, so safety
is not a concern here.

Introduces new structs representing the RESULT:Rows CQL frame body,
in various phases of deserialization:
- `RawMetadataAndRawRows` only deserializes (flags and) paging size
  in order to pass it directly to the user; keeps metadata and rows
  in a serialized, unparsed form;
- `DeserializedMetadataAndRawRows` deserializes metadata, but keeps
  rows serialized.

Credits to Karol Baryła for replacing a macro (the first approach)
with a generic function.
Even bigger credits to Karol Baryła for finding yoke crate and using it
to simplify design and empower us with flexible self-borrowing metadata
containers!

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>
Co-authored-by: Karol Baryła <karol.baryla@scylladb.com>