session: make generic and introduce "session kind" parameter
The LegacySession and the upcoming Session will differ on a small
number of methods, but otherwise will share remaining ones. In order to
reduce boilerplate the (Legacy)Session is converted into a generic,
with a type parameter indicating the kind of the API it supports
(legacy or the current one). The common methods will be implemented for
GenericSession<K> for any K, and methods specific to the particular kind
will only be implemented for GenericSession<K> for that particular K.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>