session: add get_keyspace() method
Adds the `get_keyspace()` method which allows to get the keyspace name
currently set for the session. Initially, it is set to the keyspace name
specified in the session configuration as the keyspace to be USE-d, but
it will be updated when calling `Session::use_keyspace()` later. Because
of that possibility, `get_keyspace` returns a shared pointer to the
keyspace name instead of a reference.