treewide: appease clippy
The usual: clippy have been updated in GitHub's CI runner image that we
use and it has new lints that our codebase doesn't currently pass.

This time, clippy complains about some `impl Default`s being
unnecessarily hand-written where they could be easily implemented with
an automatic derive. This commit removes the problematic impl for
PartitionerName and replaces it with a derive, but the impl for
Consistency is left as is and marked with an "allow" annotation.
Unfortunately, implementing Default with a derive requires adding a
\#[default] annotation in the enum body which incorrectly interacts with
num_enum::TryFromPrimitive macro.