load_balancing: allow to disable shuffling replicas
When token awareness is enabled in the default load balancing policy,
it chooses a random replica to send the query to. This is supposed
to balance the load however in some scenarios it might be more
beneficial to always choose the same replica - for example, reads might
benefit from better caching. This commit adds an option to disable this
behavior, which will cause the load balancing policy to always choose
nodes in the same order for each replica set.

This feature was inspired by similar functionality in the cpp driver.