load_balancing: introduce ExpectedGroup
Currently, the ExpectedGroups type verifies load balancing plans like
this: it splits the plan into multiple groups of lengths specified by
the ExpectedGroups object, and then verifies that each group has the
correct set of nodes (nodes within a single group can be in any order).

In the future we would like to check different kinds of assertions, so
the ExpectedGroup enum (notice the singular form) is introduced. It will
allow checking different kinds of assertions on a group than the current
one, but - for now - it only verifies that the set of nodes in a single
group is correct.