load_balancing: verify non-determinism in the LBP
As an additional validation which checks that the default load balancing
shuffles nodes if and only if it should do so, the ExpectedGroups struct
now also checks that the non-deterministic groups appear to be shuffled
in different plans.

There is a chance that all plans generated in a test will all have the
same order in a non-deterministic group, failing the check even if it
actually works. In order to reduce the flakiness to minimum, the number
of created load balancing policies is increased from 16 to 256. I ran
the test 1000 times and didn't get a failure (previously it occurred
quite frequently).