session: refactor to calculate_token_and_partition_key
`Session::calculate_token` calculates a serialized partition key in
order to calculate the token, but we would like to include both in a
tracing::Span. In order to be able to do that, most of the function is
extracted to `calculate_token_and_routing_key` and the old, public
method is reimplemented with the new function.