{session,iterator}: support for iter methods
Now, like non-iterator methods, the iterators also create RequestSpans
and fill them up properly.

Due to specifics of how the control flow is organized in the iterator,
it's not really possible to create RequestSpans that would cover a
successful request and predecessing, failed attempts (which should be
the case for non-iterator methods). A separate RequestSpan will be
created for each failed attempt. In order to fix that, a major refactor
of the iterator worker will have to be done.