pull_request_template.md: add reminder to update docs
It's important to have documentation that is up-to-date, and the best
time to update it is along with the changes introduced to the code. As
not everybody remembers to do it every time (me, in particular), let's
add a pre-review check about introducing docstrings and adjusting the
documentation in the book.