docs: don't strip Rust attributes
The `prepare_sphinx_source.py` script goes over snippets of Rust code
in the docs and removes the lines that start with a pound sign (#).
This is done to simulate the mdbook behavior: the snippets are supposed
to be valid Rust code that compiles, however making them compile may
require adding some boilerplate lines of code that shouldn't be
displayed in the final documentation. Sphinx, used to generate the
official documentation, doesn't understand that so we need a script to
pre-process.

However, the current behavior of the script is not entirely correct. Not
every line that starts with a pound sign should be removed - for
example, `derive` attributes like this are currently stripped but
shouldn't:

    #[derive(Debug)]
    struct Foo {
        bar: i32,
    }

In order to fix this, the code is adjusted to strip lines that start
with the '# ' string (pound sign followed by a space). This is what the
`rustdoc` tool actually does (and `mdbook`, too). This results in
correctly pre-processed snippets of code in the documentation generated
by sphinx.