iterator: make token awareness work again with execute_iter
Due to an oversight, when the RowIterator asks the load balancing policy
to calculate the plan (sequence of nodes that will be tried), it passes
a RoutingInfo structure without providing the keyspace relevant to the
query. The old token aware policy used to fall back to SimpleStrategy
and rf=1, which could provide an illusion of token-aware routing.
However, the new default load balancing policy will, in the case
described above, treat this as if no routing information is available
and will choose a random node.

This commit fixes the simple issue and adds a test which verifies that
token awareness works both in case of `execute` and `execute_iter`.