session_builder: known_nodes accept IntoIterator
There is no reason why not accept arbitrary impl IntoIterator in
SessionBuilder::[known_nodes(), known_nodes_addr()]. It will also save
an allocation in bindings.

Note that the changes in usage (removing references from arrays) are for
clippy only ("the borrowed expression implements the requested traits").
The new API is backwards-compatible. To assert that, a compile-time
check was added. It verifies that any value satisfing the old
constraints satisfies the new constraints as well.