session_test: fix flakiness due to slow Cassandra tracing
The driver's default configuration for fetching traces from the DB is
quite unforgiving. After the query finishes, it tries 5 times to fetch
tracing information, waiting 3ms between each attempt. This is too fast
for Cassandra in CI and the `test_token_awareness` fails due to that.

This commit modifies the test so that it uses a much more relaxed
configuration: 50 attempts with 200ms sleeps in between. As a long term
solution, we should consider changing the default parameters.