default_policy: introduce ReplicaLocationPreference
As the case of providing preferred rack without specifying the preferred
datacenter is invalid, in spirit of the principle “Make illegal states
unrepresentable”, an enum is introduced to restrict the possible cases.
So far, it is only used in DefaultPolicy, but in the next major release
DefaultBuilder's API should be changed accordingly to disallow creating
such configuration.

It appears that there was a bug which involved incorrect behaviour when
preferred rack was provided without setting preferred datacenter:
instead on ignoring the preferred rack, replicas were filtered by it.
Now, the following is true:
it is no longer possible to turn on rack awareness without dc-awareness;
attempts to do so will result in fallback to non-rack awareness.