default_policy: LA tests: don't rely on timing
The test bearing an undeniably attractive name: `latency_aware_default_
policy_stops_penalising_after_min_average_increases_enough_only_after_
update_rate_elapses` has shown to be flaky. I suspect it is due to the
fact that it relies on timing to be valid, and execution in CI is often
extremely slow. The problem, most probably, lies in minimum average
updates done periodically: the test code assumes that an update has not
yet been triggered at some point, which in CI happens to not be true.
To combat that, those tests are changed to use the MinAvgUpdater
explicitly instead of scheduling its periodic runs. This prevents
flakiness.