CI: add RUST_BACKTRACE=full env variable to the workflows
Currently, if a test panics in CI, a panic message will be printed which
does not contain too much of the context. For example:

    thread 'transport::session_test::test_tracing' panicked at 'No rows for tracing with this session id!', scylla/src/transport/session_test.rs:1093:5

From this information, it's not always possible to learn much about the
issue. In this particular case, the `test_tracing` case calls the
function which panicked from multiple places and it's impossible to tell
which place has this problem specifically.

To help with debugging such issues, this commit modifies the workflow
definitions to set the `RUST_BACKTRACE` environment variable. This will
make sure that the panic will print a full backtrace from the place
where this panic originated from, which should help narrow down the code
path that failed.